requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
sentence-transformers>=2.2.0

# Unsloth and TRL for efficient training
unsloth
trl
vllm

# Additional dependencies
numpy>=1.24.0
pynvml
bitsandbytes
xformers
triton

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0
matplotlib>=3.7.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Jupyter for notebook support
jupyter>=1.0.0
ipykernel>=6.25.0